m10_12_01/pyproject.toml
[tool.poetry]
name = "m10-12-01"
version = "0.1.0"
description = ""
authors = ["Yuriy Kuchma <krabatua@gmail.com>"]
readme = "README.md"
packages = [{include = "m10_12_01"}]

[tool.poetry.dependencies]
python = "^3.10"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
fastapi = "^0.95.1"
uvicorn = "^0.22.0"
sqlalchemy = "^2.0.12"
python-multipart = "^0.0.6"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
